# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * e-MarketBridge API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterInputRole =
  (typeof RegisterInputRole)[keyof typeof RegisterInputRole];

export const RegisterInputRole = {
  admin: "admin",
  buyer: "buyer",
} as const;

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  buyer: "buyer",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  storeId?: number | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserUpdate {
  name?: string;
  email?: string;
}

export interface Store {
  id: number;
  userId: number;
  storeName: string;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  description?: string | null;
  location: string;
  ruralArea: string;
  /** @nullable */
  contactInfo?: string | null;
  createdAt: string;
}

export interface StoreInput {
  /** @minLength 1 */
  storeName: string;
  logo?: string;
  description?: string;
  location: string;
  ruralArea: string;
  contactInfo?: string;
}

export interface StoreUpdate {
  storeName?: string;
  logo?: string;
  description?: string;
  location?: string;
  ruralArea?: string;
  contactInfo?: string;
}

export interface Product {
  id: number;
  storeId: number;
  /** @nullable */
  storeName?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  stock: number;
  category: string;
  /** @nullable */
  images?: string | null;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  stock: number;
  category: string;
  images?: string;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 0 */
  stock?: number;
  category?: string;
  images?: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  /** @nullable */
  productName?: string | null;
  quantity: number;
  price: number;
}

export interface Order {
  id: number;
  buyerId: number;
  /** @nullable */
  buyerName?: string | null;
  totalAmount: number;
  status: OrderStatus;
  shippingAddress: string;
  items?: OrderItem[];
  createdAt: string;
}

export type OrderInputPaymentMethod =
  (typeof OrderInputPaymentMethod)[keyof typeof OrderInputPaymentMethod];

export const OrderInputPaymentMethod = {
  cash_on_delivery: "cash_on_delivery",
  card: "card",
} as const;

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderInput {
  /** @minLength 1 */
  shippingAddress: string;
  paymentMethod: OrderInputPaymentMethod;
  items: OrderItemInput[];
}

export type OrderStatusUpdateStatus =
  (typeof OrderStatusUpdateStatus)[keyof typeof OrderStatusUpdateStatus];

export const OrderStatusUpdateStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export type PaymentPaymentMethod =
  (typeof PaymentPaymentMethod)[keyof typeof PaymentPaymentMethod];

export const PaymentPaymentMethod = {
  cash_on_delivery: "cash_on_delivery",
  card: "card",
} as const;

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface Payment {
  id: number;
  orderId: number;
  paymentMethod: PaymentPaymentMethod;
  /** @nullable */
  transactionId?: string | null;
  status: PaymentStatus;
  createdAt: string;
}

export type PaymentInputPaymentMethod =
  (typeof PaymentInputPaymentMethod)[keyof typeof PaymentInputPaymentMethod];

export const PaymentInputPaymentMethod = {
  cash_on_delivery: "cash_on_delivery",
  card: "card",
} as const;

export interface PaymentInput {
  orderId: number;
  paymentMethod: PaymentInputPaymentMethod;
  cardNumber?: string;
  cardExpiry?: string;
  cardCvv?: string;
}

export interface DashboardStats {
  totalSales: number;
  totalOrders: number;
  totalProducts: number;
  totalRevenue: number;
  pendingOrders: number;
}

export interface MonthlySales {
  month: string;
  sales: number;
  revenue: number;
}

export interface TopProduct {
  productId: number;
  name: string;
  totalSold: number;
  revenue: number;
}

export type ListProductsParams = {
  category?: string;
  storeId?: number;
  search?: string;
};