# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: e-MarketBridge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # USERS
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (buyer sees own, admin sees store orders)
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # PAYMENTS
  /payments:
    post:
      operationId: processPayment
      tags: [payments]
      summary: Process payment for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{orderId}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ANALYTICS (admin)
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [admin, buyer]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # USERS
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, buyer]
        storeId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    Order:
      type: object
      required: [id, buyerId, totalAmount, status, shippingAddress, createdAt]
      properties:
        id:
          type: integer
        buyerId:
          type: integer
        buyerName:
          type: ["string", "null"]
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        shippingAddress:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      required: [id, orderId, productId, quantity, price]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: number
    OrderInput:
      type: object
      required: [shippingAddress, items, paymentMethod]
      properties:
        shippingAddress:
          type: string
          minLength: 1
        paymentMethod:
          type: string
          enum: [cash_on_delivery, card]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]

    # PAYMENTS
    Payment:
      type: object
      required: [id, orderId, paymentMethod, status, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        paymentMethod:
          type: string
          enum: [cash_on_delivery, card]
        transactionId:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed, failed]
        createdAt:
          type: string
          format: date-time
    PaymentInput:
      type: object
      required: [orderId, paymentMethod]
      properties:
        orderId:
          type: integer
        paymentMethod:
          type: string
          enum: [cash_on_delivery, card]
        cardNumber:
          type: string
        cardExpiry:
          type: string
        cardCvv:
          type: string

    # ANALYTICS
...